Crate google_monitoring3

This documentation was generated from Monitoring crate version 1.0.9+20190629, where 20190629 is the exact revision of the monitoring:v3 schema built by the mako code generator v1.0.9.

Everything else about the Monitoring v3 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

projects

alert policies create, alert policies delete, alert policies get, alert policies list, alert policies patch, collectd time series create, groups create, groups delete, groups get, groups list, groups members list, groups update, metric descriptors create, metric descriptors delete, metric descriptors get, metric descriptors list, monitored resource descriptors get, monitored resource descriptors list, notification channel descriptors get, notification channel descriptors list, notification channels create, notification channels delete, notification channels get, notification channels get verification code, notification channels list, notification channels patch, notification channels send verification code, notification channels verify, time series create, time series list, uptime check configs create, uptime check configs delete, uptime check configs get, uptime check configs list and uptime check configs patch

uptime check ips

list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().metric_descriptors_delete(...).doit()
let r = hub.projects().alert_policies_delete(...).doit()
let r = hub.projects().uptime_check_configs_delete(...).doit()
let r = hub.projects().notification_channels_delete(...).doit()
let r = hub.projects().time_series_create(...).doit()
let r = hub.projects().notification_channels_send_verification_code(...).doit()
let r = hub.projects().groups_delete(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-monitoring3 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_monitoring3 as monitoring3;
use monitoring3::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use monitoring3::Monitoring;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Monitoring::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().notification_channels_delete("name")
             .force(true)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Aggregation
Describes how to combine multiple time series to provide different views of the data. Aggregation consists of an alignment step on individual time series (alignment_period and per_series_aligner) followed by an optional reduction step of the data across the aligned time series (cross_series_reducer and group_by_fields). For more details, see Aggregation.

AlertPolicy
A description of the conditions under which some aspect of your system is considered to be "unhealthy" and the ways to notify people or services about this state. For an overview of alert policies, see Introduction to Alerting.

BasicAuthentication
A type of authentication to perform against the specified resource or URL that uses username and password. Currently, only Basic authentication is supported in Uptime Monitoring.

BucketOptions
BucketOptions describes the bucket boundaries used to create a histogram for the distribution. The buckets can be in a linear sequence, an exponential sequence, or each bucket can be specified explicitly. BucketOptions does not include the number of values in each bucket.A bucket has an inclusive lower bound and exclusive upper bound for the values that are counted for that bucket. The upper bound of a bucket must be strictly greater than the lower bound. The sequence of N buckets for a distribution consists of an underflow bucket (number 0), zero or more finite buckets (number 1 through N - 2) and an overflow bucket (number N - 1). The buckets are contiguous: the lower bound of bucket i (i > 0) is the same as the upper bound of bucket i - 1. The buckets span the whole range of finite values: lower bound of the underflow bucket is -infinity and the upper bound of the overflow bucket is +infinity. The finite buckets are so-called because both bounds are finite.

Chunk

CollectdPayload
A collection of data points sent from a collectd-based plugin. See the collectd documentation for more information.

CollectdPayloadError
Describes the error status for payloads that were not written.

CollectdValue
A single data point from a collectd-based plugin.

CollectdValueError
Describes the error status for values that were not written.

Condition
A condition is a true/false test that determines when an alerting policy should open an incident. If a condition evaluates to true, it signifies that something is wrong.

ContentMatcher
Used to perform string matching. It allows substring and regular expressions, together with their negations.

ContentRange
Implements the Content-Range header, for serialization only

CreateCollectdTimeSeriesRequest
The CreateCollectdTimeSeries request.

CreateCollectdTimeSeriesResponse
The CreateCollectdTimeSeries response.

CreateTimeSeriesRequest
The CreateTimeSeries request.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Distribution
Distribution contains summary statistics for a population of values. It optionally contains a histogram representing the distribution of those values across a set of buckets.The summary statistics are the count, mean, sum of the squared deviation from the mean, the minimum, and the maximum of the set of population of values. The histogram is based on a sequence of buckets and gives a count of values that fall into each bucket. The boundaries of the buckets are given either explicitly or by formulas for buckets of fixed or exponentially increasing widths.Although it is not forbidden, it is generally a bad idea to include non-finite values (infinities or NaNs) in the population of values, as this will render the mean and sum_of_squared_deviation fields meaningless.

Documentation
A content string and a MIME type that describes the content string's format.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

Exemplar
Exemplars are example points that may be used to annotate aggregated distribution values. They are metadata that gives information about a particular value added to a Distribution bucket, such as a trace ID that was active when a value was added. They may contain further information, such as a example values and timestamps, origin, etc.

Explicit
Specifies a set of buckets with arbitrary widths.There are size(bounds) + 1 (= N) buckets. Bucket i has the following boundaries:Upper bound (0 <= i < N-1): boundsi Lower bound (1 <= i < N); boundsi - 1The bounds field must contain at least one element. If bounds has only one element, then there are no finite buckets, and that single element is the common boundary of the overflow and underflow buckets.

Exponential
Specifies an exponential sequence of buckets that have a width that is proportional to the value of the lower bound. Each bucket represents a constant relative uncertainty on a specific value in the bucket.There are num_finite_buckets + 2 (= N) buckets. Bucket i has the following boundaries:Upper bound (0 <= i < N-1): scale * (growth_factor ^ i). Lower bound (1 <= i < N): scale * (growth_factor ^ (i - 1)).

GetNotificationChannelVerificationCodeRequest
The GetNotificationChannelVerificationCode request.

GetNotificationChannelVerificationCodeResponse
The GetNotificationChannelVerificationCode request.

Group
The description of a dynamic collection of monitored resources. Each group has a filter that is matched against monitored resources and their associated metadata. If a group's filter matches an available monitored resource, then that resource is a member of that group. Groups can contain any number of monitored resources, and each monitored resource can be a member of any number of groups.Groups can be nested in parent-child hierarchies. The parentName field identifies an optional parent for each group. If a group has a parent, then the only monitored resources available to be matched by the group's filter are the resources contained in the parent group. In other words, a group contains the monitored resources that match its filter and the filters of all the group's ancestors. A group without a parent can contain any monitored resource.For example, consider an infrastructure running a set of instances with two user-defined tags: "environment" and "role". A parent group has a filter, environment="production". A child of that parent group has a filter, role="transcoder". The parent group contains all instances in the production environment, regardless of their roles. The child group contains instances that have the transcoder role and are in the production environment.The monitored resources contained in a group can change at any moment, depending on what resources exist and what filters are associated with the group and its ancestors.

HttpCheck
Information involved in an HTTP/HTTPS uptime check request.

InternalChecker
An internal checker allows uptime checks to run on private/internal GCP resources.

JsonServerError
A utility type which can decode a server response that indicates error

LabelDescriptor
A description of a label.

Linear
Specifies a linear sequence of buckets that all have the same width (except overflow and underflow). Each bucket represents a constant absolute uncertainty on the specific value in the bucket.There are num_finite_buckets + 2 (= N) buckets. Bucket i has the following boundaries:Upper bound (0 <= i < N-1): offset + (width * i). Lower bound (1 <= i < N): offset + (width * (i - 1)).

ListAlertPoliciesResponse
The protocol for the ListAlertPolicies response.

ListGroupMembersResponse
The ListGroupMembers response.

ListGroupsResponse
The ListGroups response.

ListMetricDescriptorsResponse
The ListMetricDescriptors response.

ListMonitoredResourceDescriptorsResponse
The ListMonitoredResourceDescriptors response.

ListNotificationChannelDescriptorsResponse
The ListNotificationChannelDescriptors response.

ListNotificationChannelsResponse
The ListNotificationChannels response.

ListTimeSeriesResponse
The ListTimeSeries response.

ListUptimeCheckConfigsResponse
The protocol for the ListUptimeCheckConfigs response.

ListUptimeCheckIpsResponse
The protocol for the ListUptimeCheckIps response.

MethodInfo
Contains information about an API request.

Metric
A specific metric, identified by specifying values for all of the labels of a MetricDescriptor.

MetricAbsence
A condition type that checks that monitored resources are reporting data. The configuration defines a metric and a set of monitored resources. The predicate is considered in violation when a time series for the specified metric of a monitored resource does not include any data in the specified duration.

MetricDescriptor
Defines a metric type and its schema. Once a metric descriptor is created, deleting or altering it stops data collection and makes the metric type's existing data unusable.

MetricDescriptorMetadata
Additional annotations that can be used to guide the usage of a metric.

MetricThreshold
A condition type that compares a collection of time series against a threshold.

MonitoredResource
An object representing a resource that can be used for monitoring, logging, billing, or other purposes. Examples include virtual machine instances, databases, and storage devices such as disks. The type field identifies a MonitoredResourceDescriptor object that describes the resource's schema. Information in the labels field identifies the actual resource and its attributes according to the schema. For example, a particular Compute Engine VM instance could be represented by the following object, because the MonitoredResourceDescriptor for "gce_instance" has labels "instance_id" and "zone": { "type": "gce_instance", "labels": { "instance_id": "12345678901234", "zone": "us-central1-a" }}

MonitoredResourceDescriptor
An object that describes the schema of a MonitoredResource object using a type name and a set of labels. For example, the monitored resource descriptor for Google Compute Engine VM instances has a type of "gce_instance" and specifies the use of the labels "instance_id" and "zone" to identify particular VM instances.Different APIs can support different monitored resource types. APIs generally provide a list method that returns the monitored resource descriptors used by the API.

MonitoredResourceMetadata
Auxiliary metadata for a MonitoredResource object. MonitoredResource objects contain the minimum set of information to uniquely identify a monitored resource instance. There is some other useful auxiliary metadata. Monitoring and Logging use an ingestion pipeline to extract metadata for cloud resources of all types, and store the metadata in this message.

Monitoring
Central instance to access all Monitoring related resource activities

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

MutationRecord
Describes a change made to a configuration.

NotificationChannel
A NotificationChannel is a medium through which an alert is delivered when a policy violation is detected. Examples of channels include email, SMS, and third-party messaging applications. Fields containing sensitive information like authentication tokens or contact info are only partially populated on retrieval.

NotificationChannelDescriptor
A description of a notification channel. The descriptor includes the properties of the channel and the set of labels or fields that must be specified to configure channels of a given type.

Point
A single data point in a time series.

ProjectAlertPolicyCreateCall
Creates a new alerting policy.

ProjectAlertPolicyDeleteCall
Deletes an alerting policy.

ProjectAlertPolicyGetCall
Gets a single alerting policy.

ProjectAlertPolicyListCall
Lists the existing alerting policies for the project.

ProjectAlertPolicyPatchCall
Updates an alerting policy. You can either replace the entire policy with a new one or replace only certain fields in the current alerting policy by specifying the fields to be updated via updateMask. Returns the updated alerting policy.

ProjectCollectdTimeSeryCreateCall
Stackdriver Monitoring Agent only: Creates a new time series.

ProjectGroupCreateCall
Creates a new group.

ProjectGroupDeleteCall
Deletes an existing group.

ProjectGroupGetCall
Gets a single group.

ProjectGroupListCall
Lists the existing groups.

ProjectGroupMemberListCall
Lists the monitored resources that are members of a group.

ProjectGroupUpdateCall
Updates an existing group. You can change any group attributes except name.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Monitoring hub.

ProjectMetricDescriptorCreateCall
Creates a new metric descriptor. User-created metric descriptors define custom metrics.

ProjectMetricDescriptorDeleteCall
Deletes a metric descriptor. Only user-created custom metrics can be deleted.

ProjectMetricDescriptorGetCall
Gets a single metric descriptor. This method does not require a Stackdriver account.

ProjectMetricDescriptorListCall
Lists metric descriptors that match a filter. This method does not require a Stackdriver account.

ProjectMonitoredResourceDescriptorGetCall
Gets a single monitored resource descriptor. This method does not require a Stackdriver account.

ProjectMonitoredResourceDescriptorListCall
Lists monitored resource descriptors that match a filter. This method does not require a Stackdriver account.

ProjectNotificationChannelCreateCall
Creates a new notification channel, representing a single notification endpoint such as an email address, SMS number, or PagerDuty service.

ProjectNotificationChannelDeleteCall
Deletes a notification channel.

ProjectNotificationChannelDescriptorGetCall
Gets a single channel descriptor. The descriptor indicates which fields are expected / permitted for a notification channel of the given type.

ProjectNotificationChannelDescriptorListCall
Lists the descriptors for supported channel types. The use of descriptors makes it possible for new channel types to be dynamically added.

ProjectNotificationChannelGetCall
Gets a single notification channel. The channel includes the relevant configuration details with which the channel was created. However, the response may truncate or omit passwords, API keys, or other private key matter and thus the response may not be 100% identical to the information that was supplied in the call to the create method.

ProjectNotificationChannelGetVerificationCodeCall
Requests a verification code for an already verified channel that can then be used in a call to VerifyNotificationChannel() on a different channel with an equivalent identity in the same or in a different project. This makes it possible to copy a channel between projects without requiring manual reverification of the channel. If the channel is not in the verified state, this method will fail (in other words, this may only be used if the SendNotificationChannelVerificationCode and VerifyNotificationChannel paths have already been used to put the given channel into the verified state).There is no guarantee that the verification codes returned by this method will be of a similar structure or form as the ones that are delivered to the channel via SendNotificationChannelVerificationCode; while VerifyNotificationChannel() will recognize both the codes delivered via SendNotificationChannelVerificationCode() and returned from GetNotificationChannelVerificationCode(), it is typically the case that the verification codes delivered via SendNotificationChannelVerificationCode() will be shorter and also have a shorter expiration (e.g. codes such as "G-123456") whereas GetVerificationCode() will typically return a much longer, websafe base 64 encoded string that has a longer expiration time.

ProjectNotificationChannelListCall
Lists the notification channels that have been created for the project.

ProjectNotificationChannelPatchCall
Updates a notification channel. Fields not specified in the field mask remain unchanged.

ProjectNotificationChannelSendVerificationCodeCall
Causes a verification code to be delivered to the channel. The code can then be supplied in VerifyNotificationChannel to verify the channel.

ProjectNotificationChannelVerifyCall
Verifies a NotificationChannel by proving receipt of the code delivered to the channel as a result of calling SendNotificationChannelVerificationCode.

ProjectTimeSeryCreateCall
Creates or adds data to one or more time series. The response is empty if all time series in the request were written. If any time series could not be written, a corresponding failure message is included in the error response.

ProjectTimeSeryListCall
Lists time series that match a filter. This method does not require a Stackdriver account.

ProjectUptimeCheckConfigCreateCall
Creates a new uptime check configuration.

ProjectUptimeCheckConfigDeleteCall
Deletes an uptime check configuration. Note that this method will fail if the uptime check configuration is referenced by an alert policy or other dependent configs that would be rendered invalid by the deletion.

ProjectUptimeCheckConfigGetCall
Gets a single uptime check configuration.

ProjectUptimeCheckConfigListCall
Lists the existing valid uptime check configurations for the project, leaving out any invalid configurations.

ProjectUptimeCheckConfigPatchCall
Updates an uptime check configuration. You can either replace the entire configuration with a new one or replace only certain fields in the current configuration by specifying the fields to be updated via "updateMask". Returns the updated configuration.

Range
The range of the population values.

RangeResponseHeader

ResourceGroup
The resource submessage for group checks. It can be used instead of a monitored resource, when multiple resources are being monitored.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

SendNotificationChannelVerificationCodeRequest
The SendNotificationChannelVerificationCode request.

ServerError

ServerMessage

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC (https://github.com/grpc). Each Status message contains three pieces of data: error code, error message, and error details.You can find out more about this error model and how to work with it in the API Design Guide (https://cloud.google.com/apis/design/errors).

TcpCheck
Information required for a TCP uptime check request.

TimeInterval
A closed time interval. It extends from the start time to the end time, and includes both: [startTime, endTime]. Valid time intervals depend on the MetricKind of the metric value. In no case can the end time be earlier than the start time. For a GAUGE metric, the startTime value is technically optional; if no value is specified, the start time defaults to the value of the end time, and the interval represents a single point in time. Such an interval is valid only for GAUGE metrics, which are point-in-time measurements. For DELTA and CUMULATIVE metrics, the start time must be later than the end time. In all cases, the start time of the next interval must be at least a microsecond after the end time of the previous interval. Because the interval is closed, if the start time of a new interval is the same as the end time of the previous interval, data written at the new start time could overwrite data written at the previous end time.

TimeSeries
A collection of data points that describes the time-varying values of a metric. A time series is identified by a combination of a fully-specified monitored resource and a fully-specified metric. This type is used for both listing and creating time series.

Trigger
Specifies how many time series must fail a predicate to trigger a condition. If not specified, then a {count: 1} trigger is used.

TypedValue
A single strongly-typed value.

UptimeCheckConfig
This message configures which resources and services to monitor for availability.

UptimeCheckIp
Contains the region, location, and list of IP addresses where checkers in the location run from.

UptimeCheckIpListCall
Returns the list of IPs that checkers run from

UptimeCheckIpMethods
A builder providing access to all methods supported on uptimeCheckIp resources. It is not used directly, but through the Monitoring hub.

VerifyNotificationChannelRequest
The VerifyNotificationChannel request.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.