Struct google_monitoring3::api::FolderTimeSeryListCall

pub struct FolderTimeSeryListCall<'a, C> where
    C: 'a,  { /* fields omitted */ }

Lists time series that match a filter. This method does not require a Workspace.

A builder for the timeSeries.list method supported by a folder resource. It is not used directly, but through a FolderMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.folders().time_series_list("name")
             .view("eos")
             .secondary_aggregation_per_series_aligner("dolor")
             .add_secondary_aggregation_group_by_fields("ea")
             .secondary_aggregation_cross_series_reducer("ipsum")
             .secondary_aggregation_alignment_period("invidunt")
             .page_token("amet")
             .page_size(-20)
             .order_by("ipsum")
             .interval_start_time("sed")
             .interval_end_time("ut")
             .filter("gubergren")
             .aggregation_per_series_aligner("rebum.")
             .add_aggregation_group_by_fields("est")
             .aggregation_cross_series_reducer("ipsum")
             .aggregation_alignment_period("ipsum")
             .doit().await;

Implementations

impl<'a, C> FolderTimeSeryListCall<'a, C> where
C: BorrowMut<Client<HttpsConnector<HttpConnector>, Body>>,

pub async fn doit(self) -> Result<(Response<Body>, ListTimeSeriesResponse)>

Perform the operation you have build so far.

pub fn name(self, new_value: &str) -> FolderTimeSeryListCall<'a, C>

Required. The project, organization or folder on which to execute the request. The format is: projects/[PROJECT_ID_OR_NUMBER] organizations/[ORGANIZATION_ID] folders/[FOLDER_ID]

Sets the name path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn view(self, new_value: &str) -> FolderTimeSeryListCall<'a, C>

Required. Specifies which information is returned about the time series.

Sets the view query property to the given value.

pub fn secondary_aggregation_per_series_aligner(
self,
new_value: &str
) -> FolderTimeSeryListCall<'a, C>

An Aligner describes how to bring the data points in a single time series into temporal alignment. Except for ALIGN_NONE, all alignments cause all the data points in an alignment_period to be mathematically grouped together, resulting in a single data point for each alignment_period with end timestamp at the end of the period.Not all alignment operations may be applied to all time series. The valid choices depend on the metric_kind and value_type of the original time series. Alignment can change the metric_kind or the value_type of the time series.Time series data must be aligned in order to perform cross-time series reduction. If cross_series_reducer is specified, then per_series_aligner must be specified and not equal to ALIGN_NONE and alignment_period must be specified; otherwise, an error is returned.

Sets the secondary aggregation.per series aligner query property to the given value.

pub fn add_secondary_aggregation_group_by_fields(
self,
new_value: &str
) -> FolderTimeSeryListCall<'a, C>

The set of fields to preserve when cross_series_reducer is specified. The group_by_fields determine how the time series are partitioned into subsets prior to applying the aggregation operation. Each subset contains time series that have the same value for each of the grouping fields. Each individual time series is a member of exactly one subset. The cross_series_reducer is applied to each subset of time series. It is not possible to reduce across different resource types, so this field implicitly contains resource.type. Fields not specified in group_by_fields are aggregated away. If group_by_fields is not specified and all the time series have the same resource type, then the time series are aggregated into a single output time series. If cross_series_reducer is not defined, this field is ignored.

Append the given value to the secondary aggregation.group by fields query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

pub fn secondary_aggregation_cross_series_reducer(
self,
new_value: &str
) -> FolderTimeSeryListCall<'a, C>

The reduction operation to be used to combine time series into a single time series, where the value of each data point in the resulting series is a function of all the already aligned values in the input time series.Not all reducer operations can be applied to all time series. The valid choices depend on the metric_kind and the value_type of the original time series. Reduction can yield a time series with a different metric_kind or value_type than the input time series.Time series data must first be aligned (see per_series_aligner) in order to perform cross-time series reduction. If cross_series_reducer is specified, then per_series_aligner must be specified, and must not be ALIGN_NONE. An alignment_period must also be specified; otherwise, an error is returned.

Sets the secondary aggregation.cross series reducer query property to the given value.

pub fn secondary_aggregation_alignment_period(
self,
new_value: &str
) -> FolderTimeSeryListCall<'a, C>

The alignment_period specifies a time interval, in seconds, that is used to divide the data in all the time series into consistent blocks of time. This will be done before the per-series aligner can be applied to the data.The value must be at least 60 seconds. If a per-series aligner other than ALIGN_NONE is specified, this field is required or an error is returned. If no per-series aligner is specified, or the aligner ALIGN_NONE is specified, then this field is ignored.The maximum value of the alignment_period is 104 weeks (2 years) for charts, and 90,000 seconds (25 hours) for alerting policies.

Sets the secondary aggregation.alignment period query property to the given value.

pub fn page_token(self, new_value: &str) -> FolderTimeSeryListCall<'a, C>

If this field is not empty then it must contain the nextPageToken value returned by a previous call to this method. Using this field causes the method to return additional results from the previous method call.

Sets the page token query property to the given value.

pub fn page_size(self, new_value: i32) -> FolderTimeSeryListCall<'a, C>

A positive number that is the maximum number of results to return. If page_size is empty or more than 100,000 results, the effective page_size is 100,000 results. If view is set to FULL, this is the maximum number of Points returned. If view is set to HEADERS, this is the maximum number of TimeSeries returned.

Sets the page size query property to the given value.

pub fn order_by(self, new_value: &str) -> FolderTimeSeryListCall<'a, C>

Unsupported: must be left blank. The points in each time series are currently returned in reverse time order (most recent to oldest).

Sets the order by query property to the given value.

pub fn interval_start_time(
self,
new_value: &str
) -> FolderTimeSeryListCall<'a, C>

Optional. The beginning of the time interval. The default value for the start time is the end time. The start time must not be later than the end time.

Sets the interval.start time query property to the given value.

pub fn interval_end_time(self, new_value: &str) -> FolderTimeSeryListCall<'a, C>

Required. The end of the time interval.

Sets the interval.end time query property to the given value.

pub fn filter(self, new_value: &str) -> FolderTimeSeryListCall<'a, C>

Required. A monitoring filter (https://cloud.google.com/monitoring/api/v3/filters) that specifies which time series should be returned. The filter must specify a single metric type, and can additionally specify metric labels and other information. For example: metric.type = "compute.googleapis.com/instance/cpu/usage_time" AND metric.labels.instance_name = "my-instance-name"

Sets the filter query property to the given value.

pub fn aggregation_per_series_aligner(
self,
new_value: &str
) -> FolderTimeSeryListCall<'a, C>

An Aligner describes how to bring the data points in a single time series into temporal alignment. Except for ALIGN_NONE, all alignments cause all the data points in an alignment_period to be mathematically grouped together, resulting in a single data point for each alignment_period with end timestamp at the end of the period.Not all alignment operations may be applied to all time series. The valid choices depend on the metric_kind and value_type of the original time series. Alignment can change the metric_kind or the value_type of the time series.Time series data must be aligned in order to perform cross-time series reduction. If cross_series_reducer is specified, then per_series_aligner must be specified and not equal to ALIGN_NONE and alignment_period must be specified; otherwise, an error is returned.

Sets the aggregation.per series aligner query property to the given value.

pub fn add_aggregation_group_by_fields(
self,
new_value: &str
) -> FolderTimeSeryListCall<'a, C>

The set of fields to preserve when cross_series_reducer is specified. The group_by_fields determine how the time series are partitioned into subsets prior to applying the aggregation operation. Each subset contains time series that have the same value for each of the grouping fields. Each individual time series is a member of exactly one subset. The cross_series_reducer is applied to each subset of time series. It is not possible to reduce across different resource types, so this field implicitly contains resource.type. Fields not specified in group_by_fields are aggregated away. If group_by_fields is not specified and all the time series have the same resource type, then the time series are aggregated into a single output time series. If cross_series_reducer is not defined, this field is ignored.

Append the given value to the aggregation.group by fields query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

pub fn aggregation_cross_series_reducer(
self,
new_value: &str
) -> FolderTimeSeryListCall<'a, C>

The reduction operation to be used to combine time series into a single time series, where the value of each data point in the resulting series is a function of all the already aligned values in the input time series.Not all reducer operations can be applied to all time series. The valid choices depend on the metric_kind and the value_type of the original time series. Reduction can yield a time series with a different metric_kind or value_type than the input time series.Time series data must first be aligned (see per_series_aligner) in order to perform cross-time series reduction. If cross_series_reducer is specified, then per_series_aligner must be specified, and must not be ALIGN_NONE. An alignment_period must also be specified; otherwise, an error is returned.

Sets the aggregation.cross series reducer query property to the given value.

pub fn aggregation_alignment_period(
self,
new_value: &str
) -> FolderTimeSeryListCall<'a, C>

The alignment_period specifies a time interval, in seconds, that is used to divide the data in all the time series into consistent blocks of time. This will be done before the per-series aligner can be applied to the data.The value must be at least 60 seconds. If a per-series aligner other than ALIGN_NONE is specified, this field is required or an error is returned. If no per-series aligner is specified, or the aligner ALIGN_NONE is specified, then this field is ignored.The maximum value of the alignment_period is 104 weeks (2 years) for charts, and 90,000 seconds (25 hours) for alerting policies.

Sets the aggregation.alignment period query property to the given value.

pub fn delegate(
self,
new_value: &'a mut dyn Delegate
) -> FolderTimeSeryListCall<'a, C>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(self, name: T, value: T) -> FolderTimeSeryListCall<'a, C> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

$.xgafv (query-string) - V1 error format.

access_token (query-string) - OAuth access token.

alt (query-string) - Data format for response.

callback (query-string) - JSONP

fields (query-string) - Selector specifying which fields to include in a partial response.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

oauth_token (query-string) - OAuth 2.0 token for the current user.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").

pub fn add_scope<T, S>(self, scope: T) -> FolderTimeSeryListCall<'a, C> where
T: Into<Option<S>>,
S: AsRef<str>,

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::CloudPlatform.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope. If None is specified, then all scopes will be removed and no default scope will be used either. In that case, you have to specify your API-key using the key parameter (see the param() function for details).

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.