Struct google_monitoring3::Distribution

pub struct Distribution {
    pub count: Option<String>,
    pub bucket_counts: Option<Vec<i64>>,
    pub sum_of_squared_deviation: Option<f64>,
    pub range: Option<Range>,
    pub bucket_options: Option<BucketOptions>,
    pub mean: Option<f64>,
}

Distribution contains summary statistics for a population of values and, optionally, a histogram representing the distribution of those values across a specified set of histogram buckets. The summary statistics are the count, mean, sum of the squared deviation from the mean, the minimum, and the maximum of the set of population of values. The histogram is based on a sequence of buckets and gives a count of values that fall into each bucket. The boundaries of the buckets are given either explicitly or by specifying parameters for a method of computing them (buckets of fixed width or buckets of exponentially increasing width). Although it is not forbidden, it is generally a bad idea to include non-finite values (infinities or NaNs) in the population of values, as this will render the mean and sum_of_squared_deviation fields meaningless.

This type is not used in any activity, and only used as part of another schema.

Fields

count

The number of values in the population. Must be non-negative.

bucket_counts

If bucket_options is given, then the sum of the values in bucket_counts must equal the value in count. If bucket_options is not given, no bucket_counts fields may be given. Bucket counts are given in order under the numbering scheme described above (the underflow bucket has number 0; the finite buckets, if any, have numbers 1 through N-2; the overflow bucket has number N-1). The size of bucket_counts must be no greater than N as defined in bucket_options. Any suffix of trailing zero bucket_count fields may be omitted.

sum_of_squared_deviation

The sum of squared deviations from the mean of the values in the population. For values x_i this is: Sumi=1..n Knuth, "The Art of Computer Programming", Vol. 2, page 323, 3rd edition describes Welford's method for accumulating this sum in one pass. If count is zero then this field must be zero.

range

If specified, contains the range of the population values. The field must not be present if the count is zero.

bucket_options

Defines the histogram bucket boundaries.

mean

The arithmetic mean of the values in the population. If count is zero then this field must be zero.