Struct google_monitoring3::Explicit

pub struct Explicit {
    pub bounds: Option<Vec<f64>>,
}

A set of buckets with arbitrary widths. Defines size(bounds) + 1 (= N) buckets with these boundaries for bucket i: Upper bound (0 <= i < N-1): bounds[i] Lower bound (1 <= i < N); bounds[i - 1] There must be at least one element in bounds. If bounds has only one element, there are no finite buckets, and that single element is the common boundary of the overflow and underflow buckets.

This type is not used in any activity, and only used as part of another schema.

Fields
bounds: Option<Vec<f64>>

The values must be monotonically increasing.