Struct google_monitoring3::MetricDescriptor

pub struct MetricDescriptor {
    pub display_name: Option<String>,
    pub name: Option<String>,
    pub metric_kind: Option<String>,
    pub value_type: Option<String>,
    pub labels: Option<Vec<LabelDescriptor>>,
    pub type_: Option<String>,
    pub unit: Option<String>,
    pub description: Option<String>,
}

Defines a metric type and its schema.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

metric descriptors get projects (response)

metric descriptors create projects (request|response)

Fields

display_name

A concise name for the metric, which can be displayed in user interfaces. Use sentence case without an ending period, for example "Request count".

name

Resource name. The format of the name may vary between different implementations. For examples: projects/{project_id}/metricDescriptors/{type=**} metricDescriptors/{type=**}

metric_kind

Whether the metric records instantaneous values, changes to a value, etc.

value_type

Whether the measurement is an integer, a floating-point number, etc.

labels

The set of labels that can be used to describe a specific instance of this metric type. For example, the compute.googleapis.com/instance/network/received_bytes_count metric type has a label, loadbalanced, that specifies whether the traffic was received through a load balanced IP address.

type_

The metric type including a DNS name prefix, for example "compute.googleapis.com/instance/cpu/utilization". Metric types should use a natural hierarchical grouping such as the following: compute.googleapis.com/instance/cpu/utilization compute.googleapis.com/instance/disk/read_ops_count compute.googleapis.com/instance/network/received_bytes_count Note that if the metric type changes, the monitoring data will be discontinued, and anything depends on it will break, such as monitoring dashboards, alerting rules and quota limits. Therefore, once a metric has been published, its type should be immutable.

unit

The unit in which the metric value is reported. It is only applicable if the value_type is INT64, DOUBLE, or DISTRIBUTION. The supported units are a subset of The Unified Code for Units of Measure standard: Basic units (UNIT) * bit bit * By byte * s second * min minute * h hour * d day Prefixes (PREFIX) * k kilo (10**3) * M mega (10**6) * G giga (10**9) * T tera (10**12) * P peta (10**15) * E exa (10**18) * Z zetta (10**21) * Y yotta (10**24) * m milli (10**-3) * u micro (10**-6) * n nano (10**-9) * p pico (10**-12) * f femto (10**-15) * a atto (10**-18) * z zepto (10**-21) * y yocto (10**-24) * Ki kibi (2**10) * Mi mebi (2**20) * Gi gibi (2**30) * Ti tebi (2**40) Grammar The grammar includes the dimensionless unit 1, such as 1/s. The grammar also includes these connectors: * / division (as an infix operator, e.g. 1/s). * . multiplication (as an infix operator, e.g. GBy.d) The grammar for a unit is as follows: Expression = Component { "." Component } { "/" Component } ; Component = [ PREFIX ] UNIT [ Annotation ] | Annotation | "1" ; Annotation = "{" NAME "}" ; Notes: * Annotation is just a comment if it follows a UNIT and is equivalent to 1 if it is used alone. For examples, {requests}/s == 1/s, By{transmitted}/s == By/s. * NAME is a sequence of non-blank printable ASCII characters not containing '{' or '}'.

description

A detailed description of the metric, which can be used in documentation.