Struct google_monitoring3::MonitoredResourceDescriptor

pub struct MonitoredResourceDescriptor {
    pub type_: Option<String>,
    pub labels: Option<Vec<LabelDescriptor>>,
    pub display_name: Option<String>,
    pub name: Option<String>,
    pub description: Option<String>,
}

An object that describes the schema of a MonitoredResource object using a type name and a set of labels. For example, the monitored resource descriptor for Google Compute Engine VM instances has a type of "gce_instance" and specifies the use of the labels "instance_id" and "zone" to identify particular VM instances. Different APIs can support different monitored resource types. APIs generally provide a list method that returns the monitored resource descriptors used by the API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

monitored resource descriptors get projects (response)

Fields
type_: Option<String>

Required. The monitored resource type. For example, the type "cloudsql_database" represents databases in Google Cloud SQL.

labels: Option<Vec<LabelDescriptor>>

Required. A set of labels used to describe instances of this monitored resource type. For example, an individual Google Cloud SQL database is identified by values for the labels "database_id" and "zone".

display_name: Option<String>

Optional. A concise name for the monitored resource type that might be displayed in user interfaces. For example, "Google Cloud SQL Database".

name: Option<String>

Optional. The resource name of the monitored resource descriptor: "projects/{project_id}/monitoredResourceDescriptors/{type}" where {type} is the value of the type field in this object and {project_id} is a project ID that provides API-specific context for accessing the type. APIs that do not use project information can use the resource name format "monitoredResourceDescriptors/{type}".

description: Option<String>

Optional. A detailed description of the monitored resource type that might be used in documentation.