Struct google_monitoring3::ProjectMethods

pub struct ProjectMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Monitoring hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_monitoring3 as monitoring3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use monitoring3::Monitoring;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Monitoring::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `alert_policies_create(...)`, `alert_policies_delete(...)`, `alert_policies_get(...)`, `alert_policies_list(...)`, `alert_policies_patch(...)`, `collectd_time_series_create(...)`, `groups_create(...)`, `groups_delete(...)`, `groups_get(...)`, `groups_list(...)`, `groups_members_list(...)`, `groups_update(...)`, `metric_descriptors_create(...)`, `metric_descriptors_delete(...)`, `metric_descriptors_get(...)`, `metric_descriptors_list(...)`, `monitored_resource_descriptors_get(...)`, `monitored_resource_descriptors_list(...)`, `notification_channel_descriptors_get(...)`, `notification_channel_descriptors_list(...)`, `notification_channels_create(...)`, `notification_channels_delete(...)`, `notification_channels_get(...)`, `notification_channels_get_verification_code(...)`, `notification_channels_list(...)`, `notification_channels_patch(...)`, `notification_channels_send_verification_code(...)`, `notification_channels_verify(...)`, `time_series_create(...)`, `time_series_list(...)`, `uptime_check_configs_create(...)`, `uptime_check_configs_delete(...)`, `uptime_check_configs_get(...)`, `uptime_check_configs_list(...)` and `uptime_check_configs_patch(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

pub fn notification_channels_get_verification_code(
&self,
request: GetNotificationChannelVerificationCodeRequest,
name: &str
) -> ProjectNotificationChannelGetVerificationCodeCall<'a, C, A>

Create a builder to help you perform the following task:

Requests a verification code for an already verified channel that can then be used in a call to VerifyNotificationChannel() on a different channel with an equivalent identity in the same or in a different project. This makes it possible to copy a channel between projects without requiring manual reverification of the channel. If the channel is not in the verified state, this method will fail (in other words, this may only be used if the SendNotificationChannelVerificationCode and VerifyNotificationChannel paths have already been used to put the given channel into the verified state).There is no guarantee that the verification codes returned by this method will be of a similar structure or form as the ones that are delivered to the channel via SendNotificationChannelVerificationCode; while VerifyNotificationChannel() will recognize both the codes delivered via SendNotificationChannelVerificationCode() and returned from GetNotificationChannelVerificationCode(), it is typically the case that the verification codes delivered via SendNotificationChannelVerificationCode() will be shorter and also have a shorter expiration (e.g. codes such as "G-123456") whereas GetVerificationCode() will typically return a much longer, websafe base 64 encoded string that has a longer expiration time.

Arguments

request - No description provided.

name - The notification channel for which a verification code is to be generated and retrieved. This must name a channel that is already verified; if the specified channel is not verified, the request will fail.

pub fn notification_channels_get(
&self,
name: &str
) -> ProjectNotificationChannelGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single notification channel. The channel includes the relevant configuration details with which the channel was created. However, the response may truncate or omit passwords, API keys, or other private key matter and thus the response may not be 100% identical to the information that was supplied in the call to the create method.

Arguments

name - The channel for which to execute the request. The format is projects/[PROJECT_ID]/notificationChannels/[CHANNEL_ID].

pub fn metric_descriptors_get(
&self,
name: &str
) -> ProjectMetricDescriptorGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single metric descriptor. This method does not require a Stackdriver account.

Arguments

name - The metric descriptor on which to execute the request. The format is "projects/{project_id_or_number}/metricDescriptors/{metric_id}". An example value of {metric_id} is "compute.googleapis.com/instance/disk/read_bytes_count".

pub fn notification_channels_create(
&self,
request: NotificationChannel,
name: &str
) -> ProjectNotificationChannelCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new notification channel, representing a single notification endpoint such as an email address, SMS number, or PagerDuty service.

Arguments

request - No description provided.

name - The project on which to execute the request. The format is: projects/[PROJECT_ID] Note that this names the container into which the channel will be written. This does not name the newly created channel. The resulting channel's name will have a normalized version of this field as a prefix, but will add /notificationChannels/[CHANNEL_ID] to identify the channel.

pub fn notification_channels_patch(
&self,
request: NotificationChannel,
name: &str
) -> ProjectNotificationChannelPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a notification channel. Fields not specified in the field mask remain unchanged.

Arguments

request - No description provided.

name - The full REST resource name for this channel. The syntax is: projects/[PROJECT_ID]/notificationChannels/[CHANNEL_ID] The [CHANNEL_ID] is automatically assigned by the server on creation.

pub fn groups_list(&self, name: &str) -> ProjectGroupListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the existing groups.

Arguments

name - The project whose groups are to be listed. The format is "projects/{project_id_or_number}".

pub fn groups_create(
&self,
request: Group,
name: &str
) -> ProjectGroupCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new group.

Arguments

request - No description provided.

name - The project in which to create the group. The format is "projects/{project_id_or_number}".

pub fn groups_get(&self, name: &str) -> ProjectGroupGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single group.

Arguments

name - The group to retrieve. The format is "projects/{project_id_or_number}/groups/{group_id}".

pub fn notification_channel_descriptors_list(
&self,
name: &str
) -> ProjectNotificationChannelDescriptorListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the descriptors for supported channel types. The use of descriptors makes it possible for new channel types to be dynamically added.

Arguments

name - The REST resource name of the parent from which to retrieve the notification channel descriptors. The expected syntax is: projects/[PROJECT_ID] Note that this names the parent container in which to look for the descriptors; to retrieve a single descriptor by name, use the GetNotificationChannelDescriptor operation, instead.

pub fn time_series_create(
&self,
request: CreateTimeSeriesRequest,
name: &str
) -> ProjectTimeSeryCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates or adds data to one or more time series. The response is empty if all time series in the request were written. If any time series could not be written, a corresponding failure message is included in the error response.

Arguments

request - No description provided.

name - The project on which to execute the request. The format is "projects/{project_id_or_number}".

pub fn metric_descriptors_list(
&self,
name: &str
) -> ProjectMetricDescriptorListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists metric descriptors that match a filter. This method does not require a Stackdriver account.

Arguments

name - The project on which to execute the request. The format is "projects/{project_id_or_number}".

pub fn monitored_resource_descriptors_list(
&self,
name: &str
) -> ProjectMonitoredResourceDescriptorListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists monitored resource descriptors that match a filter. This method does not require a Stackdriver account.

Arguments

name - The project on which to execute the request. The format is "projects/{project_id_or_number}".

pub fn uptime_check_configs_get(
&self,
name: &str
) -> ProjectUptimeCheckConfigGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single uptime check configuration.

Arguments

name - The uptime check configuration to retrieve. The format is projects/[PROJECT_ID]/uptimeCheckConfigs/[UPTIME_CHECK_ID].

pub fn notification_channel_descriptors_get(
&self,
name: &str
) -> ProjectNotificationChannelDescriptorGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single channel descriptor. The descriptor indicates which fields are expected / permitted for a notification channel of the given type.

Arguments

name - The channel type for which to execute the request. The format is projects/[PROJECT_ID]/notificationChannelDescriptors/{channel_type}.

pub fn time_series_list(&self, name: &str) -> ProjectTimeSeryListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists time series that match a filter. This method does not require a Stackdriver account.

Arguments

name - The project on which to execute the request. The format is "projects/{project_id_or_number}".

pub fn alert_policies_list(
&self,
name: &str
) -> ProjectAlertPolicyListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the existing alerting policies for the project.

Arguments

name - The project whose alert policies are to be listed. The format is projects/[PROJECT_ID] Note that this field names the parent container in which the alerting policies to be listed are stored. To retrieve a single alerting policy by name, use the GetAlertPolicy operation, instead.

pub fn metric_descriptors_delete(
&self,
name: &str
) -> ProjectMetricDescriptorDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a metric descriptor. Only user-created custom metrics can be deleted.

Arguments

name - The metric descriptor on which to execute the request. The format is "projects/{project_id_or_number}/metricDescriptors/{metric_id}". An example of {metric_id} is: "custom.googleapis.com/my_test_metric".

pub fn collectd_time_series_create(
&self,
request: CreateCollectdTimeSeriesRequest,
name: &str
) -> ProjectCollectdTimeSeryCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Stackdriver Monitoring Agent only: Creates a new time series.

Arguments

request - No description provided.

name - The project in which to create the time series. The format is "projects/PROJECT_ID_OR_NUMBER".

pub fn alert_policies_create(
&self,
request: AlertPolicy,
name: &str
) -> ProjectAlertPolicyCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new alerting policy.

Arguments

request - No description provided.

name - The project in which to create the alerting policy. The format is projects/[PROJECT_ID].Note that this field names the parent container in which the alerting policy will be written, not the name of the created policy. The alerting policy that is returned will have a name that contains a normalized representation of this name as a prefix but adds a suffix of the form /alertPolicies/[POLICY_ID], identifying the policy in the container.

pub fn alert_policies_patch(
&self,
request: AlertPolicy,
name: &str
) -> ProjectAlertPolicyPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an alerting policy. You can either replace the entire policy with a new one or replace only certain fields in the current alerting policy by specifying the fields to be updated via updateMask. Returns the updated alerting policy.

Arguments

request - No description provided.

name - Required if the policy exists. The resource name for this policy. The syntax is: projects/[PROJECT_ID]/alertPolicies/[ALERT_POLICY_ID] [ALERT_POLICY_ID] is assigned by Stackdriver Monitoring when the policy is created. When calling the alertPolicies.create method, do not include the name field in the alerting policy passed as part of the request.

pub fn notification_channels_list(
&self,
name: &str
) -> ProjectNotificationChannelListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the notification channels that have been created for the project.

Arguments

name - The project on which to execute the request. The format is projects/[PROJECT_ID]. That is, this names the container in which to look for the notification channels; it does not name a specific channel. To query a specific channel by REST resource name, use the GetNotificationChannel operation.

pub fn groups_update(
&self,
request: Group,
name: &str
) -> ProjectGroupUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing group. You can change any group attributes except name.

Arguments

request - No description provided.

name - Output only. The name of this group. The format is "projects/{project_id_or_number}/groups/{group_id}". When creating a group, this field is ignored and a new name is created consisting of the project specified in the call to CreateGroup and a unique {group_id} that is generated automatically.

pub fn notification_channels_send_verification_code(
&self,
request: SendNotificationChannelVerificationCodeRequest,
name: &str
) -> ProjectNotificationChannelSendVerificationCodeCall<'a, C, A>

Create a builder to help you perform the following task:

Causes a verification code to be delivered to the channel. The code can then be supplied in VerifyNotificationChannel to verify the channel.

Arguments

request - No description provided.

name - The notification channel to which to send a verification code.

pub fn uptime_check_configs_patch(
&self,
request: UptimeCheckConfig,
name: &str
) -> ProjectUptimeCheckConfigPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an uptime check configuration. You can either replace the entire configuration with a new one or replace only certain fields in the current configuration by specifying the fields to be updated via "updateMask". Returns the updated configuration.

Arguments

request - No description provided.

name - A unique resource name for this UptimeCheckConfig. The format is:projects/[PROJECT_ID]/uptimeCheckConfigs/[UPTIME_CHECK_ID].This field should be omitted when creating the uptime check configuration; on create, the resource name is assigned by the server and included in the response.

pub fn uptime_check_configs_create(
&self,
request: UptimeCheckConfig,
parent: &str
) -> ProjectUptimeCheckConfigCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new uptime check configuration.

Arguments

request - No description provided.

parent - The project in which to create the uptime check. The format is projects/[PROJECT_ID].

pub fn alert_policies_delete(
&self,
name: &str
) -> ProjectAlertPolicyDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an alerting policy.

Arguments

name - The alerting policy to delete. The format is: projects/[PROJECT_ID]/alertPolicies/[ALERT_POLICY_ID] For more information, see AlertPolicy.

pub fn uptime_check_configs_delete(
&self,
name: &str
) -> ProjectUptimeCheckConfigDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an uptime check configuration. Note that this method will fail if the uptime check configuration is referenced by an alert policy or other dependent configs that would be rendered invalid by the deletion.

Arguments

name - The uptime check configuration to delete. The format is projects/[PROJECT_ID]/uptimeCheckConfigs/[UPTIME_CHECK_ID].

pub fn notification_channels_delete(
&self,
name: &str
) -> ProjectNotificationChannelDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a notification channel.

Arguments

name - The channel for which to execute the request. The format is projects/[PROJECT_ID]/notificationChannels/[CHANNEL_ID].

pub fn groups_members_list(
&self,
name: &str
) -> ProjectGroupMemberListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the monitored resources that are members of a group.

Arguments

name - The group whose members are listed. The format is "projects/{project_id_or_number}/groups/{group_id}".

pub fn alert_policies_get(
&self,
name: &str
) -> ProjectAlertPolicyGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single alerting policy.

Arguments

name - The alerting policy to retrieve. The format is projects/[PROJECT_ID]/alertPolicies/[ALERT_POLICY_ID]

pub fn metric_descriptors_create(
&self,
request: MetricDescriptor,
name: &str
) -> ProjectMetricDescriptorCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new metric descriptor. User-created metric descriptors define custom metrics.

Arguments

request - No description provided.

name - The project on which to execute the request. The format is "projects/{project_id_or_number}".

pub fn groups_delete(&self, name: &str) -> ProjectGroupDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an existing group.

Arguments

name - The group to delete. The format is "projects/{project_id_or_number}/groups/{group_id}".

pub fn uptime_check_configs_list(
&self,
parent: &str
) -> ProjectUptimeCheckConfigListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the existing valid uptime check configurations for the project, leaving out any invalid configurations.

Arguments

parent - The project whose uptime check configurations are listed. The format is projects/[PROJECT_ID].

pub fn monitored_resource_descriptors_get(
&self,
name: &str
) -> ProjectMonitoredResourceDescriptorGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single monitored resource descriptor. This method does not require a Stackdriver account.

Arguments

name - The monitored resource descriptor to get. The format is "projects/{project_id_or_number}/monitoredResourceDescriptors/{resource_type}". The {resource_type} is a predefined type, such as cloudsql_database.

pub fn notification_channels_verify(
&self,
request: VerifyNotificationChannelRequest,
name: &str
) -> ProjectNotificationChannelVerifyCall<'a, C, A>

Create a builder to help you perform the following task:

Verifies a NotificationChannel by proving receipt of the code delivered to the channel as a result of calling SendNotificationChannelVerificationCode.

Arguments

request - No description provided.

name - The notification channel to verify.