Struct google_monitoring3::ProjectTimeSeryListCall

pub struct ProjectTimeSeryListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Lists time series that match a filter.

A builder for the timeSeries.list method supported by a project resource. It is not used directly, but through a ProjectMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().time_series_list("name")
             .view("sadipscing")
             .page_token("dolor")
             .page_size(-39)
             .order_by("elitr")
             .interval_start_time("amet")
             .interval_end_time("no")
             .filter("labore")
             .aggregation_per_series_aligner("eirmod")
             .add_aggregation_group_by_fields("dolore")
             .aggregation_cross_series_reducer("invidunt")
             .aggregation_alignment_period("aliquyam")
             .doit();

Methods

impl<'a, C, A> ProjectTimeSeryListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, ListTimeSeriesResponse)>

Perform the operation you have build so far.

fn name(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

The project on which to execute the request. The format is "projects/{project_id_or_number}".

Sets the name path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn view(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

Specifies which information is returned about the time series.

Sets the view query property to the given value.

fn page_token(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

If this field is not empty then it must contain the nextPageToken value returned by a previous call to this method. Using this field causes the method to return additional results from the previous method call.

Sets the page token query property to the given value.

fn page_size(self, new_value: i32) -> ProjectTimeSeryListCall<'a, C, A>

A positive number that is the maximum number of results to return. When view field sets to FULL, it limits the number of Points server will return; if view field is HEADERS, it limits the number of TimeSeries server will return.

Sets the page size query property to the given value.

fn order_by(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

Specifies the order in which the points of the time series should be returned. By default, results are not ordered. Currently, this field must be left blank.

Sets the order by query property to the given value.

fn interval_start_time(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

(optional) If omitted, the interval is a point in time, endTime. If startTime is present, it must be earlier than (less than) endTime. The interval begins after startTime—it does not include startTime.

Sets the interval.start time query property to the given value.

fn interval_end_time(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

(required) The end of the interval. The interval includes this time.

Sets the interval.end time query property to the given value.

fn filter(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

A monitoring filter that specifies which time series should be returned. The filter must specify a single metric type, and can additionally specify metric labels and other information. For example: metric.type = "compute.googleapis.com/instance/cpu/usage_time" AND metric.label.instance_name = "my-instance-name"

Sets the filter query property to the given value.

fn aggregation_per_series_aligner(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

The approach to be used to align individual time series. Not all alignment functions may be applied to all time series, depending on the metric type and value type of the original time series. Alignment may change the metric type or the value type of the time series. Time series data must be aligned in order to perform cross-time series reduction. If crossSeriesReducer is specified, then perSeriesAligner must be specified and not equal ALIGN_NONE and alignmentPeriod must be specified; otherwise, an error is returned.

Sets the aggregation.per series aligner query property to the given value.

fn add_aggregation_group_by_fields(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

The set of fields to preserve when crossSeriesReducer is specified. The groupByFields determine how the time series are partitioned into subsets prior to applying the aggregation function. Each subset contains time series that have the same value for each of the grouping fields. Each individual time series is a member of exactly one subset. The crossSeriesReducer is applied to each subset of time series. Fields not specified in groupByFields are aggregated away. If groupByFields is not specified, the time series are aggregated into a single output time series. If crossSeriesReducer is not defined, this field is ignored.

Append the given value to the aggregation.group by fields query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn aggregation_cross_series_reducer(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

The approach to be used to combine time series. Not all reducer functions may be applied to all time series, depending on the metric type and the value type of the original time series. Reduction may change the metric type of value type of the time series. Time series data must be aligned in order to perform cross-time series reduction. If crossSeriesReducer is specified, then perSeriesAligner must be specified and not equal ALIGN_NONE and alignmentPeriod must be specified; otherwise, an error is returned.

Sets the aggregation.cross series reducer query property to the given value.

fn aggregation_alignment_period(self, new_value: &str) -> ProjectTimeSeryListCall<'a, C, A>

The alignment period for per-time series alignment. If present, alignmentPeriod must be at least 60 seconds. After per-time series alignment, each time series will contain data points only on the period boundaries. If perSeriesAligner is not specified or equals ALIGN_NONE, then this field is ignored. If perSeriesAligner is specified and does not equal ALIGN_NONE, then this field must be defined; otherwise an error is returned.

Sets the aggregation.alignment period query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> ProjectTimeSeryListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> ProjectTimeSeryListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

bearer_token (query-string) - OAuth bearer token.

pp (query-boolean) - Pretty-print response.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").

access_token (query-string) - OAuth access token.

fields (query-string) - Selector specifying which fields to include in a partial response.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

callback (query-string) - JSONP

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

alt (query-string) - Data format for response.

$.xgafv (query-string) - V1 error format.

fn add_scope<T>(self, scope: T) -> ProjectTimeSeryListCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::CloudPlatform.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.