Struct google_monitoring3::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Monitoring hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_monitoring3 as monitoring3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use monitoring3::Monitoring;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Monitoring::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `collectd_time_series_create(...)`, `groups_create(...)`, `groups_delete(...)`, `groups_get(...)`, `groups_list(...)`, `groups_members_list(...)`, `groups_update(...)`, `metric_descriptors_create(...)`, `metric_descriptors_delete(...)`, `metric_descriptors_get(...)`, `metric_descriptors_list(...)`, `monitored_resource_descriptors_get(...)`, `monitored_resource_descriptors_list(...)`, `time_series_create(...)` and `time_series_list(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn metric_descriptors_get(&self, name: &str) -> ProjectMetricDescriptorGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single metric descriptor. This method does not require a Stackdriver account.

Arguments

name - The metric descriptor on which to execute the request. The format is "projects/{project_id_or_number}/metricDescriptors/{metric_id}". An example value of {metric_id} is "compute.googleapis.com/instance/disk/read_bytes_count".

fn groups_create(&self, request: Group, name: &str) -> ProjectGroupCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new group.

Arguments

request - No description provided.

name - The project in which to create the group. The format is "projects/{project_id_or_number}".

fn groups_get(&self, name: &str) -> ProjectGroupGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single group.

Arguments

name - The group to retrieve. The format is "projects/{project_id_or_number}/groups/{group_id}".

fn metric_descriptors_delete(&self, name: &str) -> ProjectMetricDescriptorDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a metric descriptor. Only user-created custom metrics can be deleted.

Arguments

name - The metric descriptor on which to execute the request. The format is "projects/{project_id_or_number}/metricDescriptors/{metric_id}". An example of {metric_id} is: "custom.googleapis.com/my_test_metric".

fn collectd_time_series_create(&self, request: CreateCollectdTimeSeriesRequest, name: &str) -> ProjectCollectdTimeSeryCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new time series with the given data points. This method is only for use in collectd-related code, including the Google Monitoring Agent. See [google.monitoring.v3.MetricService.CreateTimeSeries] instead.

Arguments

request - No description provided.

name - The project in which to create the time series. The format is "projects/PROJECT_ID_OR_NUMBER".

fn groups_members_list(&self, name: &str) -> ProjectGroupMemberListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the monitored resources that are members of a group.

Arguments

name - The group whose members are listed. The format is "projects/{project_id_or_number}/groups/{group_id}".

fn monitored_resource_descriptors_list(&self, name: &str) -> ProjectMonitoredResourceDescriptorListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists monitored resource descriptors that match a filter. This method does not require a Stackdriver account.

Arguments

name - The project on which to execute the request. The format is "projects/{project_id_or_number}".

fn time_series_create(&self, request: CreateTimeSeriesRequest, name: &str) -> ProjectTimeSeryCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates or adds data to one or more time series. The response is empty if all time series in the request were written. If any time series could not be written, a corresponding failure message is included in the error response.

Arguments

request - No description provided.

name - The project on which to execute the request. The format is "projects/{project_id_or_number}".

fn metric_descriptors_create(&self, request: MetricDescriptor, name: &str) -> ProjectMetricDescriptorCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new metric descriptor. User-created metric descriptors define custom metrics.

Arguments

request - No description provided.

name - The project on which to execute the request. The format is "projects/{project_id_or_number}".

fn groups_list(&self, name: &str) -> ProjectGroupListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the existing groups.

Arguments

name - The project whose groups are to be listed. The format is "projects/{project_id_or_number}".

fn metric_descriptors_list(&self, name: &str) -> ProjectMetricDescriptorListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists metric descriptors that match a filter. This method does not require a Stackdriver account.

Arguments

name - The project on which to execute the request. The format is "projects/{project_id_or_number}".

fn monitored_resource_descriptors_get(&self, name: &str) -> ProjectMonitoredResourceDescriptorGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single monitored resource descriptor. This method does not require a Stackdriver account.

Arguments

name - The monitored resource descriptor to get. The format is "projects/{project_id_or_number}/monitoredResourceDescriptors/{resource_type}". The {resource_type} is a predefined type, such as cloudsql_database.

fn groups_update(&self, request: Group, name: &str) -> ProjectGroupUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing group. You can change any group attributes except name.

Arguments

request - No description provided.

name - The name of this group. The format is "projects/{project_id_or_number}/groups/{group_id}". When creating a group, this field is ignored and a new name is created consisting of the project specified in the call to CreateGroup and a unique {group_id} that is generated automatically. @OutputOnly

fn groups_delete(&self, name: &str) -> ProjectGroupDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an existing group.

Arguments

name - The group to delete. The format is "projects/{project_id_or_number}/groups/{group_id}".

fn time_series_list(&self, name: &str) -> ProjectTimeSeryListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists time series that match a filter. This method does not require a Stackdriver account.

Arguments

name - The project on which to execute the request. The format is "projects/{project_id_or_number}".